Refactor: Estructura de archivos
