+++ .gitignore
@@ -1,0 +1,2 @@
+prueba.c
+*.exe
